requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.6
contourpy==1.3.0
cycler==0.12.1
filelock==3.19.1
fonttools==4.60.2
fsspec==2025.10.0
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.12.0.88
packaging==26.0
pillow==11.3.0
polars==1.36.1
polars-runtime-32==1.36.1
psutil==7.2.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
scipy==1.13.1
six==1.17.0
sympy==1.14.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.3
typing_extensions==4.15.0
ultralytics==8.4.22
ultralytics-thop==2.0.18
urllib3==2.6.3
zipp==3.23.0
